Decimal point is changeable during display
It's not true that the decimal point in integer-based storage is hard to move. Since it implies the point must be inserted during display, and obviously, the currency sign will be added at that time, it's just fair that the currency could dictate the place the decimal point goes.
Also added some subjunctive to currency changes through history; it definitely can happen, but the text sounds like that may happen at any time, while it's actually a rare scenario.